# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    ListServicesResponse:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/PdServiceDto"
...
    PdDeploymentTarget:
      type: string
      enum:
      - ENVIRONMENT
      - PLATFORM
...
    PdServiceDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        target:
          $ref: "#/components/schemas/PdDeploymentTarget"
        branch:
          type: string
        availableOnEnv:
          type: boolean
        healthPath:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        environmentIds:
          type: array
          items:
            type: string
...
    UpsertServiceRequest:
      type: object
      properties:
        deploymentTarget:
          $ref: "#/components/schemas/PdDeploymentTarget"
        branch:
          type: string
        availableOnEnv:
          type: boolean
        healthPath:
          type: string
        environmentIds:
          type: array
          items:
            type: string
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://qits-platform-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
...
  /platform-deployments/api/services:
    get:
      summary: "Every service, with the environments each is linked into"
      responses:
        "200":
          description: The services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListServicesResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Pd Service Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
  /platform-deployments/api/services/{name}:
    put:
      summary: "Register or update one service, replacing its environment links"
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertServiceRequest"
        required: true
      responses:
        "200":
          description: The updated service
          content:
            application/json:
              schema: {}
        "201":
          description: The newly registered service
        "400":
          description: "Validation failed, or a platform service was given environment\
            \ links"
        "404":
          description: An environmentId names no environment
        "409":
          description: A platform service cannot become an environment service
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Service Controller
      security:
      - SecurityScheme:
        - qits-platform:system
    delete:
      summary: Remove a service and its links
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Removed
        "404":
          description: No such service
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Service Controller
      security:
      - SecurityScheme:
        - qits-platform:system
info:
  title: qits-platform-deployments API
  version: 2026.912.174637